apps/requirements.txt
colorthief
Pillow
